Some editorial improvements to the early sections
Added explanation of speech ambiguity and need to different speech to the accessibility section.

Improved/clarified a number of things about ARIA. In particular, some of the problems are hopefully better stated.
Added audio files to demo how an 'a' is pronounced poorly for math.

More work on CSS section including the other two examples.