Put the later two examples inside of <details> to fit the rest of the formatting.
change "segment" to "line-segment" to be more specific and also give an example where the name is speech-centric.
changed "prime" to "primed" -- seemed to be consensus that "prime" is a poor name.

also changed index because it is to generic -- how would a speech engine know to say "coordinate"?
In defence of my choice, I was able to remove the explaination of what "index" is meant to do (i.e., the example is now self-explanatory)

Added paragraph explaining that this leaves generating proper speech to AT because AT knows the preferences of the reader.