Add `by-row` and `by-column` properties (#64)
* Add `by-row` and `by-column` properties

* Add `dimensional-product`.

This was added to *both* the character list and the linear algebra section. The later is needed to indicate how it is spoken while the former is useful because that is where people will likely look for it.

It would be nice to be able to the character listing to the linear algebra one.